Create Test
In the setUp method, we create an instance of the TC class and set up the user interface components for testing.

In the testCelsiusToFahrenheitConversion method, we simulate the user input by setting the value "25" in the Celsius text field (t1), which represents 25 degrees Celsius.

We then simulate a button click on the "Convert" button (b) by calling the doClick method.

Finally, we use JUnit's assertEquals method to check if the converted value in the Fahrenheit text field (t2) matches the expected result. For the conversion from 25°C to Fahrenheit, the expected result is approximately 77.0°F.

To run this test, you can use an integrated development environment (IDE) that supports JUnit, such as Eclipse or IntelliJ IDEA. Simply right-click the TCTest class and select "Run As" -> "JUnit Test" (Eclipse) or right-click and select "Run 'TCTest'" (IntelliJ IDEA).

Make sure to have JUnit properly configured in your project for this to work.